+++ src/calculadora/Division.java
@@ -12,8 +12,8 @@ package calculadora;
 public class Division implements Calculadora{
 
     @Override
-    public float Datos(float a,float b) {
-        float c;
+     public double Datos(double a,double b) {
+        double c;
         c= a/b;
         return c;
     }
